Add more readable dates to the output
One way of interpreting these structures is to identify the earliest
point in time a snapshot of a resource was taken; equally the latest.
We can do this with the previous approach, but by extracting the
dates and parsing them correctly, we can provide more useful
information to non-technical users.